# docs/apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "基于区块链的志愿者服务认证系统 API",
    "description": "Apifox 导入文件。非公共接口统一使用 Authorization: Bearer <token>。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发环境"
    }
  ],
  "security": [{"bearerAuth": []}],
...
  "paths": {
...
    "/api/join-requests": {
      "post": {
        "tags": ["组织"],
        "summary": "申请默认归属组织",
        "description": "默认归属组织只表示志愿者档案归属，不限制活动参与范围。",
        "requestBody": {"$ref": "#/components/requestBodies/JoinOrganizationRequest"},
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
...
    "/api/join-requests/organization": {
      "get": {
        "tags": ["组织"],
        "summary": "组织归属申请列表",
        "parameters": [{"$ref": "#/components/parameters/StatusQuery"}],
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
    "/api/join-requests/{id}/review": {
      "post": {
        "tags": ["组织"],
        "summary": "审核归属组织申请",
        "parameters": [{"$ref": "#/components/parameters/IdPath"}],
        "requestBody": {"$ref": "#/components/requestBodies/ReviewJoinRequest"},
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
...
    "/api/products": {
      "post": {
        "tags": ["商城"],
        "summary": "组织创建商品草稿",
        "requestBody": {"$ref": "#/components/requestBodies/ProductSaveRequest"},
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
    "/api/products/{id}": {
      "put": {
        "tags": ["商城"],
        "summary": "组织更新商品",
        "parameters": [{"$ref": "#/components/parameters/IdPath"}],
        "requestBody": {"$ref": "#/components/requestBodies/ProductSaveRequest"},
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      },
      "delete": {
        "tags": ["商城"],
        "summary": "组织删除商品",
        "parameters": [{"$ref": "#/components/parameters/IdPath"}],
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
    "/api/products/{id}/submit": {
      "post": {
        "tags": ["商城"],
        "summary": "提交商品审核",
        "parameters": [{"$ref": "#/components/parameters/IdPath"}],
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
...
    "/api/redemptions": {
      "post": {
        "tags": ["商城"],
        "summary": "兑换商品",
        "description": "使用用户全局积分余额兑换，履约路由到商品所属组织。",
        "requestBody": {"$ref": "#/components/requestBodies/RedemptionRequest"},
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
...
    "/api/redemptions/{id}/status": {
      "post": {
        "tags": ["商城"],
        "summary": "更新兑换状态",
        "description": "fulfilled 表示已履约，cancelled 表示取消并回滚积分和库存。",
        "parameters": [{"$ref": "#/components/parameters/IdPath"}],
        "requestBody": {"$ref": "#/components/requestBodies/RedemptionStatusRequest"},
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
...
    "/api/evidences/{bizType}/{bizId}": {
      "get": {
        "tags": ["存证"],
        "summary": "登录用户查询单条存证",
        "parameters": [{"$ref": "#/components/parameters/BizTypePath"}, {"$ref": "#/components/parameters/BizIdPath"}],
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
...
    "/api/public/completed-projects": {
      "get": {
        "tags": ["公共"],
        "summary": "已完成项目展示",
        "security": [],
        "parameters": [{"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "default": 6}}],
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
    "/api/public/evidences": {
      "get": {
        "tags": ["公共"],
        "summary": "公开存证列表",
        "security": [],
        "parameters": [
          {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "default": 20}},
          {"name": "status", "in": "query", "required": false, "schema": {"type": "string"}},
          {"name": "organizationId", "in": "query", "required": false, "schema": {"type": "integer"}},
          {"name": "keyword", "in": "query", "required": false, "schema": {"type": "string"}}
        ],
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
    "/api/public/evidences/{bizType}/{bizId}": {
      "get": {
        "tags": ["公共"],
        "summary": "公开存证详情",
        "security": [],
        "parameters": [{"$ref": "#/components/parameters/BizTypePath"}, {"$ref": "#/components/parameters/BizIdPath"}],
        "responses": {"200": {"$ref": "#/components/responses/ApiResponse"}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}
    },
    "parameters": {
      "IdPath": {"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}},
...
      "BizTypePath": {"name": "bizType", "in": "path", "required": true, "schema": {"type": "string", "enum": ["completion"]}},
      "BizIdPath": {"name": "bizId", "in": "path", "required": true, "schema": {"type": "integer"}},
      "StatusQuery": {"name": "status", "in": "query", "required": false, "schema": {"type": "string"}}
    },
    "responses": {
      "ApiResponse": {
        "description": "统一响应",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ApiResponse"},
            "example": {"success": true, "message": "操作成功", "data": {}}
          }
        }
      }
    },
    "requestBodies": {
...
      "JoinOrganizationRequest": {
        "required": true,
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JoinOrganizationRequest"}, "example": {"organizationId": 1, "applyReason": "希望加入该组织。"}}}
      },
      "ReviewJoinRequest": {
        "required": true,
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReviewJoinRequest"}, "example": {"status": "approved", "reviewNote": "同意加入。"}}}
      },
...
      "ProductSaveRequest": {
        "required": true,
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductSaveRequest"}}}
      },
      "RedemptionRequest": {
        "required": true,
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RedemptionRequest"}, "example": {"productId": 1}}}
      },
      "RedemptionStatusRequest": {
        "required": true,
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RedemptionStatusRequest"}, "example": {"status": "fulfilled"}}}
      },
...
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean"},
          "message": {"type": "string"},
          "data": {
            "description": "实际类型随接口变化，字段明细见接口联调用例和后端 VO。",
            "nullable": true,
            "oneOf": [
              {"type": "object"},
              {"type": "array", "items": {}},
              {"type": "string"},
              {"type": "number"},
              {"type": "integer"},
              {"type": "boolean"}
            ]
          }
        }
      },
...
      "JoinOrganizationRequest": {
        "type": "object",
        "required": ["organizationId"],
        "properties": {
          "organizationId": {"type": "integer"},
          "applyReason": {"type": "string", "maxLength": 500}
        }
      },
      "ReviewJoinRequest": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {"type": "string", "enum": ["approved", "rejected"]},
          "reviewNote": {"type": "string", "maxLength": 500}
        }
      },
...
      "ProductSaveRequest": {
        "type": "object",
        "required": ["productName", "price", "stock"],
        "properties": {
          "productName": {"type": "string", "maxLength": 255},
          "productDescription": {"type": "string", "maxLength": 3000},
          "price": {"type": "integer", "minimum": 1},
          "stock": {"type": "integer", "minimum": 0},
          "imagePath": {"type": "string", "maxLength": 255}
        },
        "example": {
          "productName": "志愿服务纪念水杯",
          "productDescription": "统一积分商城兑换商品。",
          "price": 20,
          "stock": 100,
          "imagePath": "/uploads/products/cup.png"
        }
      },
      "RedemptionRequest": {
        "type": "object",
        "required": ["productId"],
        "properties": {"productId": {"type": "integer"}}
      },
      "RedemptionStatusRequest": {
        "type": "object",
        "required": ["status"],
        "properties": {"status": {"type": "string", "enum": ["fulfilled", "cancelled"]}}
      },
...
    }
  }
}